+++ src/main/resources/local.env.properties
@@ -5,11 +5,8 @@ AWS_DYNAMODB_BLOCK_BATCH_TABLE_NAME=Aggregate-Service-Block-Batch
 AWS_DYNAMODB_BLOCK_METADATA_TABLE_NAME=Aggregate-Service-Block-Metadata
 DYNAMODB_BATCH_GET_ITEM_MAX_ITEMS=10
 BATCH_SIZE=10
-EVENT_STREAM_WEBSOCKET_URI=http://localhost:26657
-EVENT_STREAM_RPC_URI=http://localhost:26657
 WS_NODE=http://34.148.31.174:26657
 DB_URI=http://localhost:8080
 DB_DATABASE=L2CACHE
 DB_TYPE=RAVENDB
 DB_MAX_CONNECTIONS=100
-FEE_COLLECTOR=tp17xpfvakm2amg962yls6f84z3kell8c5l2udfyt
